feat: Implement flow preview composable and store for managing flow simulation state
